system_only_extractor/requirements.txt
PyMuPDF
pandas
openpyxl
python-dotenv
openai
anthropic
